fix(skore): Document  via explicit implementation
