+++ skore/src/skore/_sklearn/_plot/metrics/roc_curve.py
@@ -11,11 +11,9 @@ from sklearn.base import BaseEstimator
 from sklearn.metrics import auc, roc_curve
 from sklearn.preprocessing import LabelBinarizer
 
-from skore._sklearn._plot.style import StyleDisplayMixin
+from skore._sklearn._plot.base import BaseDisplay
 from skore._sklearn._plot.utils import (
     LINESTYLE,
-    HelpDisplayMixin,
-    PlotBackendMixin,
     _ClassifierCurveDisplayMixin,
     _despine_matplotlib_axis,
     _validate_style_kwargs,
@@ -63,9 +61,7 @@ def _add_chance_level(
     return cast(Line2D, chance_level)
 
 
-class RocCurveDisplay(
-    StyleDisplayMixin, HelpDisplayMixin, _ClassifierCurveDisplayMixin, PlotBackendMixin
-):
+class RocCurveDisplay(_ClassifierCurveDisplayMixin, BaseDisplay):
     """ROC Curve visualization.
 
     An instance of this class should be created by `EstimatorReport.metrics.roc()`.
@@ -703,8 +699,8 @@ class RocCurveDisplay(
 
         return self.ax_, lines, info_pos_label
 
-    @StyleDisplayMixin.style_plot
-    def _plot_matplotlib(
+    @BaseDisplay.style_plot
+    def plot(
         self,
         *,
         estimator_name: str | None = None,
@@ -750,6 +746,24 @@ class RocCurveDisplay(
         >>> display = report.metrics.roc()
         >>> display.plot(roc_curve_kwargs={"color": "tab:red"})
         """
+        return self._plot(
+            estimator_name=estimator_name,
+            roc_curve_kwargs=roc_curve_kwargs,
+            plot_chance_level=plot_chance_level,
+            chance_level_kwargs=chance_level_kwargs,
+            despine=despine,
+        )
+
+    def _plot_matplotlib(
+        self,
+        *,
+        estimator_name: str | None = None,
+        roc_curve_kwargs: dict[str, Any] | list[dict[str, Any]] | None = None,
+        plot_chance_level: bool = True,
+        chance_level_kwargs: dict[str, Any] | None = None,
+        despine: bool = True,
+    ) -> None:
+        """Matplotlib implementation of the `plot` method."""
         if (
             self.report_type == "comparison-cross-validation"
             and self.ml_task == "multiclass-classification"
